Add preset for analyzing OpenBCI Ganglion data =======================================
The Ganglion has four channels instead of eight, a sample rate of 200Hz rather that 250Hz, and
has an six lines of comments before the data instead of five. This commit allows you to set the
'source' variable to 'openbci-ganglion' or 'openbci-ganglion-openvibe' to easily work with
Ganglion data. Now there's no need to edit this file before running EEGrunt on Ganglion data :)